requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
